[NEW][SERVER] Add notification in backend cache.
This is a feature lost during discussion, that we need to use different
list to distinguish timeline and notifications.

This check-in adds a cache list for each person. Right now it supports
only two notifications: follow you, or upvote your post.

The public API will be added in the following works.